Facebook iOS SDK 3.17.1
Summary:
    see https://developers.facebook.com/ios/change-log-3.x/

Upgrading:
    https://developers.facebook.com/docs/ios/upgrading/

+++ src/Login/FBSession.m
@@ -341,7 +341,7 @@ static FBSession *g_activeSession = nil;
             // valid behavior; no-op.
             break;
         default:
-            [FBLogger singleShotLogEntry:FBLoggingBehaviorDeveloperErrors formatString:@"%d is not a valid FBSessionLoginBehavior. Ignoring open call.", behavior];
+            [FBLogger singleShotLogEntry:FBLoggingBehaviorDeveloperErrors formatString:@"%lu is not a valid FBSessionLoginBehavior. Ignoring open call.", (unsigned long)behavior];
             return;
     }
     if (!(self.state == FBSessionStateCreated ||
